pyproject.toml
[tool.poetry]
name = "ocrAPI"
version = "0.1.0"
description = "Boost Camp CV 07's API practice by Jinhyung via fastAPI"
authors = ["ppjh8263 <ppjh8263@naver.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.6.2, <3.10"
fastapi = "^0.70.1"
uvicorn = "^0.16.0"
pydantic = "^1.8.2"
SQLAlchemy = "^1.4.28"
Pillow = "^8.4.0"
PyYAML = "^6.0"
requests = "^2.26.0"
python-multipart = "^0.0.5"
Jinja2 = "^3.0.3"

[tool.poetry.dev-dependencies]
poethepoet = "^0.11.0"

[tool.poe.tasks]
force-cuda11 = "python -m pip install torch==1.10.0+cu113 torchvision==0.11.1+cu113 torchaudio==0.10.0+cu113 -f https://download.pytorch.org/whl/cu113/torch_stable.html"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
    